chore: move signal impl to separate file
