+++ packages/qwik/src/core/signal/signal.ts
@@ -11,11 +11,10 @@
  *   - It needs to store a function which needs to re-run.
  *   - It is `Readonly` because it is computed.
  */
-import { isDev } from '@qwik.dev/core/build';
 import { isDomContainer } from '../client/dom-container';
 import { pad, qwikDebugToString } from '../debug';
 import type { OnRenderFn } from '../shared/component.public';
-import { assertDefined, assertFalse, assertTrue } from '../shared/error/assert';
+import { assertDefined, assertFalse } from '../shared/error/assert';
 import { QError, qError } from '../shared/error/error';
 import type { Props } from '../shared/jsx/jsx-runtime';
 import { type QRLInternal } from '../shared/qrl/qrl-class';
@@ -24,12 +23,12 @@ import type { Container, HostElement } from '../shared/types';
 import { ChoreType } from '../shared/util-chore-type';
 import { ELEMENT_PROPS, OnRenderProp } from '../shared/utils/markers';
 import { isPromise } from '../shared/utils/promises';
-import { qDev } from '../shared/utils/qdev';
 import { SerializerSymbol } from '../shared/utils/serialize-utils';
 import type { ISsrNode, SSRContainer } from '../ssr/ssr-types';
 import { trackSignal, tryGetInvokeContext } from '../use/use-core';
 import { TaskFlags, isTask } from '../use/use-task';
 import { NEEDS_COMPUTATION, _EFFECT_BACK_REF } from './flags';
+import { SignalImpl } from './impl/signal-impl';
 import { type BackRef } from './signal-cleanup';
 import type { Signal } from './signal.public';
 import type { TargetType } from './store';
@@ -65,94 +64,6 @@ export const throwIfQRLNotResolved = (qrl: QRL) => {
 export const isSignal = (value: any): value is Signal<unknown> => {
   return value instanceof SignalImpl;
 };
-
-export class SignalImpl<T = any> implements Signal<T> {
-  $untrackedValue$: T;
-
-  /** Store a list of effects which are dependent on this signal. */
-  $effects$: null | Set<EffectSubscription> = null;
-
-  $container$: Container | null = null;
-
-  constructor(container: Container | null, value: T) {
-    this.$container$ = container;
-    this.$untrackedValue$ = value;
-    DEBUG && log('new', this);
-  }
-
-  get untrackedValue() {
-    return this.$untrackedValue$;
-  }
-
-  // TODO: should we disallow setting the value directly?
-  set untrackedValue(value: T) {
-    this.$untrackedValue$ = value;
-  }
-
-  get value() {
-    const ctx = tryGetInvokeContext();
-    if (ctx) {
-      if (this.$container$ === null) {
-        if (!ctx.$container$) {
-          return this.untrackedValue;
-        }
-        // Grab the container now we have access to it
-        this.$container$ = ctx.$container$;
-      } else {
-        assertTrue(
-          !ctx.$container$ || ctx.$container$ === this.$container$,
-          'Do not use signals across containers'
-        );
-      }
-      const effectSubscriber = ctx.$effectSubscriber$;
-      if (effectSubscriber) {
-        const effects = (this.$effects$ ||= new Set());
-        // Let's make sure that we have a reference to this effect.
-        // Adding reference is essentially adding a subscription, so if the signal
-        // changes we know who to notify.
-        ensureContainsSubscription(effects, effectSubscriber);
-        // But when effect is scheduled in needs to be able to know which signals
-        // to unsubscribe from. So we need to store the reference from the effect back
-        // to this signal.
-        ensureContainsBackRef(effectSubscriber, this);
-        addQrlToSerializationCtx(effectSubscriber, this.$container$);
-        DEBUG && log('read->sub', pad('\n' + this.toString(), '  '));
-      }
-    }
-    return this.untrackedValue;
-  }
-  set value(value) {
-    if (value !== this.$untrackedValue$) {
-      DEBUG &&
-        log('Signal.set', this.$untrackedValue$, '->', value, pad('\n' + this.toString(), '  '));
-      this.$untrackedValue$ = value;
-      triggerEffects(this.$container$, this, this.$effects$);
-    }
-  }
-
-  // prevent accidental use as value
-  valueOf() {
-    if (qDev) {
-      throw qError(QError.cannotCoerceSignal);
-    }
-  }
-
-  toString() {
-    if (isDev) {
-      return (
-        `[${this.constructor.name}${(this as any).$flags$ & SignalFlags.INVALID ? ' INVALID' : ''} ${String(this.$untrackedValue$)}]` +
-        (Array.from(this.$effects$ || [])
-          .map((e) => '\n -> ' + pad(qwikDebugToString(e[0]), '    '))
-          .join('\n') || '')
-      );
-    } else {
-      return this.constructor.name;
-    }
-  }
-  toJSON() {
-    return { value: this.$untrackedValue$ };
-  }
-}
 
 export const ensureContainsSubscription = (
   array: Set<EffectSubscription>,
