+++ packages/qwik/src/core/signal/types.ts
@@ -1,10 +1,10 @@
 import type { VNode } from '../client/types';
 import type { ISsrNode } from '../ssr/ssr-types';
 import type { Task } from '../use/use-task';
-import type { SignalImpl } from './signal';
 import type { SubscriptionData } from './subscription-data';
 import type { ReadonlySignal } from './signal.public';
 import type { TargetType } from './store';
+import type { SignalImpl } from './impl/signal-impl';
 
 export interface InternalReadonlySignal<T = unknown> extends ReadonlySignal<T> {
   readonly untrackedValue: T;
